+++ debezium-quarkus-outbox-reactive/deployment/src/main/java/io/debezium/outbox/reactive/quarkus/deployment/DebeziumOutboxConfig.java
@@ -3,8 +3,9 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.outbox.quarkus.deployment;
+package io.debezium.outbox.reactive.quarkus.deployment;
 
+import io.debezium.outbox.quarkus.deployment.DebeziumOutboxCommonConfig;
 import io.quarkus.runtime.annotations.ConfigPhase;
 import io.quarkus.runtime.annotations.ConfigRoot;
