+++ debezium-quarkus-outbox-reactive/integration-tests/src/main/java/io/debezium/outbox/reactive/quarkus/it/UpperCaseAttributeConverter.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.outbox.quarkus.it;
+package io.debezium.outbox.reactive.quarkus.it;
 
 import java.util.Locale;
