+++ debezium-quarkus-outbox-reactive/integration-tests/src/main/resources/application.properties
@@ -22,4 +22,4 @@ quarkus.debezium-outbox.remove-after-insert=false
 #            this field also specifies a JPA converter to persist the value as upper-case
 # Field #3 - name_no_columndef, a java.lang.String type using ORM's default column def mappings
 #
-quarkus.debezium-outbox.additional-fields=name:string:varchar(50),name_upper:string:varchar(50):io.debezium.outbox.quarkus.it.UpperCaseAttributeConverter,name_no_columndef:string
+quarkus.debezium-outbox.additional-fields=name:string:varchar(50),name_upper:string:varchar(50):io.debezium.outbox.reactive.quarkus.it.UpperCaseAttributeConverter,name_no_columndef:string
